// fe/src/lib/api/generated/model/formRevisionStatusValue.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */

export type FormRevisionStatusValue = typeof FormRevisionStatusValue[keyof typeof FormRevisionStatusValue];


export const FormRevisionStatusValue = {
  ACTIVE: 'ACTIVE',
  INACTIVE: 'INACTIVE',
} as const;