// fe/src/lib/api/generated/model/formRevisionResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */
import type { FormRevisionStatusValue } from './formRevisionStatusValue';

export interface FormRevisionResponse {
  family_key: string;
  id: string;
  internal_schema_version: number;
  official_code: string | null;
  official_revision: string | null;
  status: FormRevisionStatusValue;
}